feat: 함수로 분리 완료
